# v1.1/backend/docs/swagger.yaml
definitions:
...
  common.DeleteRequest:
    properties:
      id:
        example: ""
        type: string
    required:
    - id
    type: object
  common.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
...
  entity.User:
    properties:
      createTime:
        type: string
      editTime:
        type: string
      id:
        type: integer
      updateTime:
        type: string
      userAccount:
        type: string
      userAvatar:
        type: string
      userName:
        type: string
      userPassword:
        type: string
      userProfile:
        type: string
      userRole:
        type: string
    type: object
...
  user.ListUserVOResponse:
    properties:
      current:
        description: 当前页数
        type: integer
      pages:
        description: 总页数
        type: integer
      records:
        items:
          $ref: '#/definitions/user.UserVO'
        type: array
      size:
        description: 页面大小
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
  user.UserAddRequest:
    properties:
      userAccount:
        description: 用户账号
        type: string
      userAvatar:
        description: 用户头像
        type: string
      userName:
        description: 用户昵称
        type: string
      userProfile:
        description: 用户简介
        type: string
      userRole:
        description: 用户权限
        type: string
    required:
    - userAccount
    type: object
  user.UserEditRequest:
    properties:
      id:
        description: 用户ID
        example: ""
        type: string
      userName:
        description: 用户昵称
        type: string
      userProfile:
        description: 用户简介
        type: string
    type: object
  user.UserLoginRequest:
    properties:
      userAccount:
        type: string
      userPassword:
        type: string
    required:
    - userAccount
    - userPassword
    type: object
  user.UserLoginVO:
    properties:
      createTime:
        type: string
      editTime:
        type: string
      id:
        example: ""
        type: string
      token:
        description: 新增Token字段
        type: string
      updateTime:
        type: string
      userAccount:
        type: string
      userAvatar:
        type: string
      userName:
        type: string
      userProfile:
        type: string
      userRole:
        type: string
    type: object
  user.UserQueryRequest:
    properties:
      current:
        description: 当前页数
        type: integer
      id:
        description: 用户ID
        example: ""
        type: string
      pageSize:
        description: 页面大小
        type: integer
      sortField:
        description: 排序字段
        type: string
      sortOrder:
        description: 排序顺序（默认升序）
        type: string
      userAccount:
        description: 用户账号
        type: string
      userName:
        description: 用户昵称
        type: string
      userProfile:
        description: 用户简介
        type: string
      userRole:
        description: 用户权限
        type: string
    type: object
  user.UserRegsiterRequest:
    properties:
      checkPassword:
        type: string
      userAccount:
        type: string
      userPassword:
        type: string
    required:
    - checkPassword
    - userAccount
    - userPassword
    type: object
  user.UserUpdateRequest:
    properties:
      id:
        description: 用户ID
        example: ""
        type: string
      userAvatar:
        description: 用户头像
        type: string
      userName:
        description: 用户昵称
        type: string
      userProfile:
        description: 用户简介
        type: string
      userRole:
        description: 用户权限
        type: string
    type: object
  user.UserVO:
    properties:
      createTime:
        type: string
      id:
        example: ""
        type: string
      userAccount:
        type: string
      userAvatar:
        type: string
      userName:
        type: string
      userProfile:
        type: string
      userRole:
        type: string
    type: object
host: localhost:8001
info:
  contact:
    email: lwh24621@qq.com
    name: lwhhhh
    url: 无
  description: CanvasCloud
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: 无服务条款
  title: CanvasCloud
  version: "1.0"
paths:
...
  /v1/user/add:
    post:
      consumes:
      - application/json
      description: 默认密码为12345678
      parameters:
      - description: 用户添加申请参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功，返回添加用户的ID
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 添加失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建一个用户「管理员」
      tags:
      - user
  /v1/user/avatar:
    post:
      consumes:
      - multipart/form-data
      description: 根据ID，将头像保存到数据库，返回是否成功
      parameters:
      - description: 图片
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功，返回图片信息视图
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 上传用户头像「需要登录校验」
      tags:
      - user
  /v1/user/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户的ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/common.DeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 删除失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 根据ID软删除用户「管理员」
      tags:
      - user
  /v1/user/edit:
    post:
      consumes:
      - application/json
      description: 若用户不存在，则返回失败
      parameters:
      - description: 需要更新的用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新用户个人资料
      tags:
      - user
  /v1/user/get:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户的ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功，返回用户的所有信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "400":
          description: 查询失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 根据ID获取用户「管理员」
      tags:
      - user
  /v1/user/get/login:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户视图成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.UserLoginVO'
              type: object
        "400":
          description: 获取失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth   // 关键：添加这个标签: []
      summary: 获取登录的用户信息
      tags:
      - user
  /v1/user/get/vo:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户的ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功，返回用户的脱敏信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.UserVO'
              type: object
        "400":
          description: 查询失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      summary: 根据ID获取简略信息用户
      tags:
      - user
  /v1/user/list/page/vo:
    post:
      consumes:
      - application/json
      description: 根据用户关键信息进行模糊查询
      parameters:
      - description: 需要查询的页数、以及用户关键信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.ListUserVOResponse'
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 分页获取一系列用户信息「管理员」
      tags:
      - user
  /v1/user/login:
    post:
      consumes:
      - application/json
      description: 根据账号密码进行登录
      parameters:
      - description: 用户登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回用户视图
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.UserLoginVO'
              type: object
        "400":
          description: 登录失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      summary: 用户登录
      tags:
      - user
  /v1/user/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: 退出成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 注册失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 执行用户注销（退出）
      tags:
      - user
  /v1/user/register:
    post:
      consumes:
      - application/json
      description: 根据账号密码进行注册
      parameters:
      - description: 用户请求注册参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserRegsiterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功，返回注册用户的ID
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 注册失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      summary: 注册用户
      tags:
      - user
  /v1/user/update:
    post:
      consumes:
      - application/json
      description: 若用户不存在，则返回失败
      parameters:
      - description: 需要更新的用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新用户信息「管理员」
      tags:
      - user
securityDefinitions:
  BearerAuth:
    description: 'JWT 认证格式: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"